requirements.txt
boto3==1.4.4              # via kinesis-python
botocore==1.5.19          # via boto3, s3transfer
click==6.7
docutils==0.13.1          # via botocore
futures==3.0.5            # via s3transfer
jmespath==0.9.1           # via boto3, botocore
kinesis-python==0.0.5
python-dateutil==2.6.0    # via botocore
s3transfer==0.1.10        # via boto3
six==1.10.0               # via python-dateutil
